requirements.txt
numpy
pandas
matplotlib
seaborn
yfinance
scikit-learn
pyyaml
tqdm
pytest
